requirement.txt
need to install:
flask
flask-cors
python-dotenv
google-adk
google-generativeai
requests
adk
axios
node.js